+++ src/collective/task/migrations/migrate_to_2_2.py
@@ -12,18 +12,14 @@ class Migrate_To_2_2(Migrator):
     def __init__(self, context):
         Migrator.__init__(self, context)
 
-    def runProfileSteps(self, profile, steps):
-        ''' Run specific steps for profile '''
-        for step_id in steps:
-            self.portal.portal_setup.runImportStepFromProfile('profile-%s:default' % profile, step_id)
-
     def run(self):
         logger.info('Migrating to collective.task 2.2')
         self.cleanRegistries()
 
-        addOrUpdateIndexes(self.portal, indexInfos={'due_date': ('DateIndex', {})})
-        addOrUpdateColumns(self.portal, columns=('due_date', ))
+        addOrUpdateIndexes(self.portal, indexInfos={'assigned_group': ('FieldIndex', {}), 'due_date': ('DateIndex', {})})
+        addOrUpdateColumns(self.portal, columns=('assigned_group', 'due_date'))
 
+        self.runProfileSteps('collective.task', steps=['workflow'])
         self.finish()