+++ src/collective/task/interfaces.py
@@ -4,14 +4,14 @@
 from zope import schema
 from zope.publisher.interfaces.browser import IDefaultBrowserLayer
 from plone.app.dexterity import PloneMessageFactory as _PMF
-from plone.dexterity.content import IDexterityContainer
+from plone.supermodel import model
 
 
 class ICollectiveTaskLayer(IDefaultBrowserLayer):
     """Marker interface that defines a browser layer."""
 
 
-class ITaskContent(IDexterityContainer):
+class ITaskContent(model.Schema):
     """ Interface for task content type """
 
     title = schema.TextLine(