+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Added due_date index. Added indexer methods avoiding acquisition for children.
   [sgeulette]
+- Changed workflow to use 'Request review' guard permission to differentiate Reviewer and editor transitions (like in plone workflows)
+  [sgeulette]
 
 
 2.1 (2015-06-30)