+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-2.2 (unreleased)
+2.2 (2015-11-24)
 ----------------
 
 - Added TaskContainer related search utility